Basic Calculator
A calculator that performs simple mathematical operations such as addition, subtraction, multiplication and division. It also checks to see whether the user has entered in a valid number/operator and if not it will send a message and prompt the user to try again, it will keep doing this until the user complies.