Timed Quiz
A two minute quiz consisting of 10 questions; each correct answer rewards the user with two points, the total of which will be displayed to them after they have completed the quiz or ran out of time. This program displays the nature of the quiz to the user at the beginning in a typewriter-like effect and it uses a thread to keep the timer running in the background for its duration.