Authentication
This project uses Python's microframework called 'Flask' to validate forms and make dynamic web pages. The program itself functions similarly to my JavaScript Form Validation v2+ projects; It includes flash messages, password hashing and a database connection to MySQL. 

It was made in a virtual environment and so the dependencies for it can be found in the requirements.txt file contained within this project folder.